Use DIY'd power measurements for all live values
